Remove code no longer used
The classes in `ca.on.oicr.gsi.provenance` were only consumed by `pipedev` and
are no longer used by Cerberus and not used by Pinery. Many of the
`ca.on.oicr.gsi.provenance.model` classes are specific to Niassa and do not map
onto Vidarr data structures, making them not useful for Cerberus or Shesmu.